Add a producer profile edition. Update sveltekit-superforms to v2 => broke some auth pages
