+++ .gitignore
@@ -11,3 +11,4 @@ vite.config.js.timestamp-*
 vite.config.ts.timestamp-*
 *.swp
 static/data/producers_*.json
+bun.lock
