+++ package.json
@@ -35,7 +35,7 @@
 		"svelte": "^5.38.10",
 		"svelte-check": "^3.6.0",
 		"sveltekit-flash-message": "^2.3.1",
-		"sveltekit-superforms": "^1.12.0",
+		"sveltekit-superforms": "^2.27.1",
 		"tailwindcss": "^3.3.2",
 		"tslib": "^2.4.1",
 		"typescript": "^5.0.0",
