+++ src/lib/server/database/drizzle.ts
@@ -1,11 +1,11 @@
 import { DATABASE_URL } from '$env/static/private';
 import { drizzle } from 'drizzle-orm/node-postgres';
 import pg from 'pg';
-const pool = new pg.Pool({
-	connectionString: DATABASE_URL
-});
+const pool = new pg.Pool({connectionString: DATABASE_URL});
 
 await pool.connect();
 const db = drizzle(pool);
+
+// console.log("DB was ", db);
 
 export default db;
